+++ NEWS.md
@@ -1,7 +1,7 @@
 # hardhat (development version)
 
-* Adds `factor_to_indicators()` function that takes a factor as input, and 
-  returns the indicators as a integer matrix.
+* New `fct_encode_one_hot()` that encodes a factor as a one-hot indicator matrix
+  (#215).
 
 * Added more documentation about importance and frequency weights in
   `?importance_weights()` and `?frequency_weights()` (#214).